complete style modification for top page
