+++ frontend/src/javascripts/app.js
@@ -9,8 +9,10 @@ import Main from './components/index/main'
 import UserOnly from './components/users/user-only'
 import GuestOnly from './components/guests/guest-only'
 import AboutPage from './components/guests/about-page'
+import TopPage from './components/top_page'
 
 import '../stylesheets/common.scss'
+import '../stylesheets/form_field.scss'
 import '../stylesheets/top_page.scss'
 import '../stylesheets/project_page.scss'
