+++ frontend/src/javascripts/components/guests/login-form.js
@@ -6,11 +6,11 @@ import { createSession } from '../../actions/users'
 
 class LoginForm extends Component {
   renderField({ placeholder, type, input, meta: { touched, error } }) {
-    const classNames = `form-group ${touched && error ? 'has-danger' : ''} fieled-text`
+    const classNames = `form-group ${touched && error ? 'has-danger' : ''} field-style fieled-text`
     return(
       <div className={classNames}>
         <input
-          className="form-control"
+          className="text-style"
           placeholder={placeholder}
           type={type}
           {...input} />
@@ -30,10 +30,9 @@ class LoginForm extends Component {
   render() {
     return(
       <form onSubmit={this.props.handleSubmit(this.onSubmit.bind(this))}>
-        <Field placeholder="EMIL ADRESS" name="email" type="text" component={this.renderField} />
-        <Field placeholder="PASSWORD" name="password" type="password" component={this.renderField} />
-        <button type="submit" className="btn btn-primary">Submit</button>
-        <Link to="/guests/signup">Sign up</Link>
+        <Field placeholder="E M A I L   A D R E S S" name="email" type="text" component={this.renderField} />
+        <Field placeholder="P A S S W O R D" name="password" type="password" component={this.renderField} />
+        <button type="submit" className="submit-btn">SIGN IN</button>
       </form>
     )
   }
