+++ frontend/src/stylesheets/_initializer.scss
@@ -1,0 +1,1 @@
+$bg-common-color: #0D171F;
