+++ frontend/src/stylesheets/common.scss
@@ -1,9 +1,21 @@
+@import "initializer";
+
+html {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+}
 body {
   width: 100%;
+  height: 100%;
+  margin: 0;
   background-image: url("../images/universe.png");
 }
-.fieled-text {
-  font-family: orator-std, monospace;
-  font-style: normal;
-  font-weight: 400;
+.container {
+  width: 100%;
+  height: 100%;
+}
+#app {
+  width: 100%;
+  height: 100%;
 }
