+++ frontend/src/stylesheets/form_field.scss
@@ -1,0 +1,74 @@
+@import "initializer";
+
+@mixin placeholder($color) {
+  &::placeholder {
+    font-size: 12px;
+    font-family: orator-std, monospace;
+    font-stretch: ultra-expanded;
+    color: $color;
+  }
+  // IE
+  &:-ms-input-placeholder {
+    font-size: 12px;
+    font-family: orator-std, monospace;
+    font-stretch: ultra-expanded;
+    color: $color;
+  }
+
+  // Edge
+  &::-ms-input-placeholder {
+    font-size: 12px;
+    font-family: orator-std, monospace;
+    font-stretch: ultra-expanded;
+    color: $color;
+  }
+
+  //mozila
+  ::-moz-placeholder {
+    font-size: 12px;
+    font-family: orator-std, monospace;
+    font-stretch: ultra-expanded;
+    color: $color;
+  }
+}
+
+.field-style {
+  display: block;
+  width: 400px;
+  margin: 20px auto;
+  color: #fff;
+
+  input {
+    @include placeholder(#808080);
+    width: 400px;
+    text-align: center;
+    background-color: $bg-common-color;
+    border: 1px solid #fff;
+    border-radius: 20px;
+    padding: 0 30px;
+  }
+}
+
+.fieled-text {
+  font-family: orator-std, monospace;
+  font-style: normal;
+  font-weight: 300;
+}
+
+.submit-btn {
+  display: block;
+  width: 200px;
+  margin: 20px auto 0;
+  padding: 1px 20px;
+  background-color: #fff;
+  border: 1px solid #fff;
+  border-radius: 20px;
+  color: $bg-common-color;
+
+  font-size: 14px;
+  line-height: 1.5;
+  font-family: sans-serif;
+  font-style: normal;
+  font-weight: 200;
+  font-stretch: ultra-expanded;
+}
